#7 Rename user to user_{directory}
